fix: add name of trab in html title and fix udpDNS to resolve to other DNS's not only ips
