+++ Modules/pedro_alan_rodrigo_henrique/frontend/index.html
@@ -15,7 +15,7 @@
   </head>
   <body>
     <div class="container">
-      <h1 style="font-size: 2rem">Packet sniffer <a href="/t2.html">Proxima</a></h1>
+      <h1 style="font-size: 2rem">T1 - Packet sniffer <a href="/t2.html">Proxima</a></h1>
       <div id="visits-per-destination" class="chart"></div>
       <div id="generated-traffic-per-source" class="chart"></div>
       <div id="visits-per-country" class="chart"></div>
