+++ Modules/pedro_alan_rodrigo_henrique/frontend/t3.html
@@ -13,7 +13,7 @@
 </head>
 <body>
   <div class="container">
-    <h2><a href="/t2.html">Anterior</a> Tabela RIP <a href="/t4.html">Próximo</a></h2>
+    <h2><a href="/t2.html">Anterior</a> T3 - Tabela RIP <a href="/t4.html">Próximo</a></h2>
     <table id="ripTable1" class="table table-striped">
       <h4>Pacotes para cada rede RIP do Roteador 1 (10.0.0.1)</h4>
       <thead>
