+++ Modules/pedro_alan_rodrigo_henrique/frontend/t5.html
@@ -15,7 +15,7 @@
   </head>
   <body>
     <div class="container">
-      <h1 style="font-size: 2rem"><a href="/t4.html">Anterior</a> TCP Analyzer</h1>
+      <h1 style="font-size: 2rem"><a href="/t4.html">Anterior</a> T5 - TCP Analyzer <a href="/t6.html">Próximo</a></h1>
       <div id="flagFrequency" class="chart"></div>
       <div id="windowSize" class="chart"></div>
       <div id="packetError" class="chart"></div>
