+++ Modules/pedro_alan_rodrigo_henrique/frontend/t6.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Análise de Pacotes UDP</title>
+  <title>Análise de pacotes HTTP</title>
   <link
     href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.1/dist/css/bootstrap.min.css"
     rel="stylesheet"
@@ -13,8 +13,8 @@
 </head>
 <body>
   <div class="container">
-    <h2><a href="/t3.html">Anterior</a> T4 - Análise pacotes UDP <a href="/t5.html">Próximo</a></h2>
+    <h2><a href="/t5.html">Anterior</a> T6 - Análise pacotes HTTP <a href="/t7.html">Próximo</a></h2>
   </div>
-  <script type="module" src="dist/t4.js"></script>
+  <script type="module" src="dist/t6.js"></script>
 </body>
 </html>
