+++ Modules/pedro_alan_rodrigo_henrique/frontend/t7.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Tabela ARP</title>
+  <title>Tabela DNS</title>
   <link
     href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.1/dist/css/bootstrap.min.css"
     rel="stylesheet"
@@ -13,24 +13,23 @@
 </head>
 <body>
   <div class="container">
-    <h2><a href="/index.html">Anterior</a> T2 - Tabela de Dispositivos <a href="/t3.html">Próxima</a></h2>
-    <table id="deviceTable" class="table table-striped">
+    <h2><a href="/t6.html">Anterior</a> T7 - Tabela de Queries DNS <a href="/t8.html">Próxima</a></h2>
+    <table id="dnsTable" class="table table-striped">
       <thead>
         <tr>
-          <th scope="col">MAC</th>
-          <th scope="col">IP</th>
-          <th scope="col">Fabricante</th>
+          <th scope="col">Domínio</th>
+          <th scope="col">Contagem</th>
+          <th scope="col">IP resolvido</th>
         </tr>
       </thead>
       <tbody> 
       </tbody>
       <div class="d-flex mb-2">
         <button onclick="start()" class="btn btn-primary me-1" id="startBtn">Start</button>
-        <input type="text" id="network" class="form-control w-50" placeholder="Network" />
-        <button class="btn btn-warning" id="setNetwork">Set network</button>
+        <button onclick="stop()" class="btn btn-danger me-1" id="stopBtn">Stop</button>
       </div>
     </table>
   </div>
-  <script type="module" src="dist/t2.js"></script>
+  <script type="module" src="dist/t7.js"></script>
 </body>
 </html>
