+++ Modules/pedro_alan_rodrigo_henrique/routers/PcapReader/PcapReader.py
@@ -9,7 +9,7 @@ class PcapReader:
 
     def __read_pcap(self):
         try:
-            self.packets = rdpcap(self.file_path)
+            self.packets = rdpcap(self.file_path, 10)
             return True
         except Exception as e:
             print("Error reading pcap file:", e)
